Add simple keyboard example (#434)
Simple example using my minimal 4x4 keymatrix. Note that `keyberon` sort
of expects a `usb_device` to be available such that it can send HID
reports. Because we don't have that (yet?), for now I'm just dumping the
data through `defmt-rtt` and, to reduce console spamming, only when the
report is not empty.

Co-authored-by: Felipe Balbi <febalbi@microsoft.com>