Use unused LUT slot in FlexSPI storage bus driver. (#439)
The storage bus driver for FlexSPI currently uses look-up-table index 0.
However, this index is also used by the ROM bootloader, and will still
be used by the CPU to access flash through the memory bus.

A simple solution is to use index 14 instead, as this is not used in the
default configuration.


![image](https://github.com/user-attachments/assets/44e9cf49-7a44-4ed4-bd79-23e567bdc0fd)

"Reserved" in the table means free to use ^_^

Note that the `mimxrt600-fcb` crate uses index 6 to configure the device
in DDR mode, so that one is also already taken.

But a user might be on a platform that overrides which command sequence
is used for what purpose, so the PR also allows the user to choose their
own.

This makes it a much easier to use the FlexSpi storage bus driver while
keeping the configuration set by the ROM bootloader from the FCB.